Trivial c2hs binding to pulseaudio
Setup Sound/Pulse/Internal.hs to import all of the raw api under
Sound/Pulse/Internal directory. And try to bind the part that is
the easiest to resolve.

Note that Sould/Pulse/Internal/C2HS is for marshallers, which at
this time are removed with C2HS module. For the details, please
check "http://stackoverflow.com/questions/9471902/whats-the-modern-way-
to-access-c2hs-marshalling-functions"